requirements.txt
aiohttp==3.12.13
evaluate==0.4.4
ipykernel==6.29.5
lxml-html-clean==0.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
peft==0.15.2
pip-chill==1.0.3
pysocks==1.7.1
rouge-score==0.1.2
sentencepiece==0.2.0
streamlit==1.46.1
torchvision==0.22.1
yfinance==0.2.64
